fix: fix sign comparison warnings in part2.cpp
Fixes a sign comparison warning by consistently using size_t for string index operations. This avoids potential issues with negative indices and integer overflow when handling string operations. The changes include:

- Convert loop variable i from int to size_t
- Convert index variable j from int to size_t
- Remove unnecessary static_cast to int in size comparisons
- Maintain identical functionality while improving type safety

The behavior of the program remains unchanged while adhering to better C++ string handling practices.

+++ day03/part2.cpp
@@ -8,7 +8,7 @@ void solve() {
   bool enabled = true;
 
   while (getline(cin, s)) {
-    for (int i = 0; i < (int)s.size() - 3; i++) {
+    for (size_t i = 0; i < s.size() - 3; i++) {
       if (s[i] == 'd' && s[i + 1] == 'o' && s[i + 2] == '(' &&
           s[i + 3] == ')') {
         enabled = true;
@@ -29,22 +29,22 @@ void solve() {
 
       if (s[i] == 'm' && s[i + 1] == 'u' && s[i + 2] == 'l' &&
           s[i + 3] == '(') {
-        int j = i + 4;
+        size_t j = i + 4;
         string x, y;
 
-        while (j < (int)s.size() && isdigit(s[j])) {
+        while (j < s.size() && isdigit(s[j])) {
           x += s[j++];
         }
 
-        if (j >= (int)s.size() || s[j] != ',' || x.size() > 3)
+        if (j >= s.size() || s[j] != ',' || x.size() > 3)
           continue;
         j++;
 
-        while (j < (int)s.size() && isdigit(s[j])) {
+        while (j < s.size() && isdigit(s[j])) {
           y += s[j++];
         }
 
-        if (j >= (int)s.size() || s[j] != ')' || y.size() > 3)
+        if (j >= s.size() || s[j] != ')' || y.size() > 3)
           continue;
 
         if (!x.empty() && !y.empty()) {
